Implement page-fnote-barsep to improve appearance
Use 2*env->fn->sep below the footnote line because the footnotes are too close
to the bar otherwise, especially when the footnote font size is the same as the
body text, as is required by some court's legal brief styles. By doubling
env->fn->sep, it looks good with either small or normal-size footnotes.